requirements.txt
astroid==1.5.2
autopep8==1.3.1
Django==1.11.2
django-extensions==1.7.8
django-modeltranslation==0.12
django-polymorphic==1.2
django-timezone-field==2.0
djangorestframework==3.6.2
isort==4.2.5
lazy-object-proxy==1.3.1
mccabe==0.6.1
pg8000==1.10.6
pycodestyle==2.3.1
pylint==1.7.1
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
testing.common.database==2.0.0
testing.postgresql==1.3.0
Werkzeug==0.12.1
wrapt==1.10.10
